Addition of new commented code, summary reports.
HMM_summary.pdf provides an overview, DOTA_InitialQuickReport.pdf goes
into more detail about the code itself and which scripts are necessary.
 RawProcessMultiCore_beforeDataTable.R is the main workhorse to
preprocess files for the HMM.  AggregateGamesCluster.R takes the
preprocessed output and generates the sequences for the HMM.
DOTA_presentation_MITRE.pptx is a presentation on the initial results.